requirements.txt
# Data manipulation and web scraping
pandas>=2.1.0,<3.0.0
lxml>=4.9.0,<5.0.0
html5lib>=1.1,<2.0

# Financial data
yfinance>=0.2.30,<1.0.0

# SEC EDGAR API
requests>=2.31.0,<3.0.0

# Database
psycopg2-binary>=2.9.9,<3.0.0

# Configuration
python-dotenv>=1.0.0,<2.0.0
